pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask-login>=0.6.3",
    "flask>=3.1.1",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "psutil>=7.0.0",
    "psycopg2-binary>=2.9.10",
    "pyyaml>=6.0.2",
    "requests>=2.32.4",
    "werkzeug>=3.1.3",
    "redis>=6.2.0",
    "celery>=5.5.3",
    "maxminddb>=2.7.0",
    "pika>=1.3.2",
    "sqlalchemy>=2.0.41",
]
